+++ addon/index.js
@@ -1,8 +1,8 @@
-import findRole from 'ralphs-little-helpers/find-role';
-import clickOn from 'ralphs-little-helpers/click-on';
-import clickRole from 'ralphs-little-helpers/click-role';
-import fillInField from 'ralphs-little-helpers/fill-in-field';
-import 'ralphs-little-helpers/extend-qunit';
+import findRole from './find-role';
+import clickOn from './click-on';
+import clickRole from './click-role';
+import fillInField from './fill-in-field';
+import './extend-qunit';
 
 export {
   clickOn,