+++ test-support/helpers/ralphs-little-helpers/test-helpers.js
@@ -1,1 +1,2 @@
+import 'ralphs-little-helpers/checkbox-helpers';
 import 'ralphs-little-helpers/within';