+++ CHANGELOG.md
@@ -2,6 +2,10 @@ master
 ------
 
 * Fix `assert.include` and `assert.notInclude` default failure message.
+* Introduce `check` and `uncheck` global async helpers.
+  Introduce `assert.checked` and `assert.unchecked`. [#9]
+
+[#9]: https://github.com/thoughtbot/ralphs-little-helpers/pull/9
 
 0.0.3
 -----